requirements.txt
python-dotenv 
langchain 
langchain-community 
langchain-openai 
langchainhub 
openai 
tiktoken 
azure-ai-documentintelligence 
azure-identity 
azure-search-documents
unstructured 